feat: day16 part 2

+++ utils/binary_to_int64.go
@@ -10,5 +10,5 @@ func BinaryToInt64(binaryNumber string) int64 {
 		panic(err)
 	}
 
-	return int64(number)
+	return number
 }
